brooklyn-storage-hazelcast: add org.apache package prefix
